infrastructure/cloudformation/pyproject.toml
[project]
name = "cloudformation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.225.0",
    "cdk-ecr-deployment>=4.0.3",
    "constructs>=10.4.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
